split rise and risf
