+++ sktime/classification/interval_based/__init__.py
@@ -11,9 +11,9 @@ __all__ = [
 
 from sktime.classification.interval_based._cif import CanonicalIntervalForest
 from sktime.classification.interval_based._drcif import DrCIF
-from sktime.classification.interval_based._rise import (
+from sktime.classification.interval_based._rise import RandomIntervalSpectralEnsemble
+from sktime.classification.interval_based._risf import (
     RandomIntervalSpectralForest,  # todo remove in 0.10.0
 )
-from sktime.classification.interval_based._rise import RandomIntervalSpectralEnsemble
 from sktime.classification.interval_based._stsf import SupervisedTimeSeriesForest
 from sktime.classification.interval_based._tsf import TimeSeriesForestClassifier
